mini-dooray-project-management-api/src/main/java/com/nhnacademy/minidoorayprojectmanagementapi/project/controller/ProjectController.java
package com.nhnacademy.minidoorayprojectmanagementapi.project.controller;

import com.nhnacademy.minidoorayprojectmanagementapi.project.dto.CreationProjectRequest;
import com.nhnacademy.minidoorayprojectmanagementapi.project.dto.ProjectCreationCompleteDto;
import com.nhnacademy.minidoorayprojectmanagementapi.project.service.ProjectService;
import org.springframework.http.HttpStatus;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.ResponseStatus;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/projects")
public class ProjectController {
    private final ProjectService projectService;

    public ProjectController(ProjectService projectService) {
        this.projectService = projectService;
    }

    @PostMapping
    @ResponseStatus(HttpStatus.CREATED)
    public ProjectCreationCompleteDto createProject(@Validated @RequestBody CreationProjectRequest projectRequest){
        return projectService.createProject(projectRequest);
    }
}
